refactor: select-search component rename
