+++ components/tarifs/index.ts
@@ -1,1 +1,3 @@
 export * from "./view";
+export * from "./constants";
+export * from "./model";
