+++ components/track-timer/view.tsx
@@ -1,7 +1,8 @@
 import { useUnit } from "effector-react";
 
 import { TIME_MODE } from "@/features/media/constants";
-import { $timeMode, $timer } from "@/src/widgets/winamp/model";
+
+import { $timeMode, $timer } from "~/widgets/winamp/model";
 
 export const TrackTimer = () => {
   const [timeMode, timer] = useUnit([$timeMode, $timer]);
