--- components/ui/stat-card/index.ts
@@ -1,1 +1,0 @@
-export * from "./view";
